Bridge native widths and strengthen cross-ABI tests
Translate CK_ULONG values and nested templates across 32-bit, 64-bit, and Windows layouts. Expand live and mock tests, then split large conversion and test modules.

+++ crates/server/src/server/grpc_service/key_ops/wrapping.rs
@@ -155,7 +155,7 @@ pub(crate) async fn unwrap_key(
                 ctx_mgr,
                 &ctx_id,
                 virtual_session,
-                CkObjectHandle(object.0),
+                CkObjectHandle(object.0 as u64),
                 is_token,
             )
             .await;
